final push

+++ Dto/BooksReviews/BookReviewDto.cs
@@ -8,6 +8,6 @@ public class BookReviewDto
     public int Id { get; set; }
     public BookMetadataDto Book { get; set; }
     public UserMetadataDto User { get; set; }
-    public int Rating { get; set; }
+    public float Rating { get; set; }
     public string Content { get; set; }
 }
